feat: 지나간 컵캣 및 하위 컴포넌트 SSR로 전환
