--- src/app/mypage/previous/_components/mock.ts
@@ -1,46 +1,0 @@
-export const cupcatData = {
-  timestamp: "2025-01-25T04:10:08.825364836",
-  code: 200,
-  message: "지나간 컵캣 정보를 성공적으로 조회했습니다.",
-  result: {
-    level: 0,
-    cupcats: [
-      {
-        cafe_name: "ccc",
-        cupcat_img_url: "https://i.ibb.co/NCyY7r1/6.png",
-      },
-      {
-        cafe_name: "ddd",
-        cupcat_img_url: "https://i.ibb.co/gS8xzBr/2-png-5.png",
-      },
-      {
-        cafe_name: "뭐시기",
-        cupcat_img_url: "https://i.ibb.co/NCyY7r1/6.png",
-      },
-      {
-        cafe_name: "eee",
-        cupcat_img_url: "https://i.ibb.co/gS8xzBr/2-png-5.png",
-      },
-      {
-        cafe_name: "fff",
-        cupcat_img_url: "https://i.ibb.co/KcPkG4c/3-6.png",
-      },
-      {
-        cafe_name: "ggg",
-        cupcat_img_url: "https://i.ibb.co/NCyY7r1/6.png",
-      },
-      {
-        cafe_name: "hhh",
-        cupcat_img_url: "https://i.ibb.co/KcPkG4c/3-6.png",
-      },
-      {
-        cafe_name: "ggg",
-        cupcat_img_url: "https://i.ibb.co/NCyY7r1/6.png",
-      },
-      {
-        cafe_name: "hhh",
-        cupcat_img_url: "https://i.ibb.co/KcPkG4c/3-6.png",
-      },
-    ],
-  },
-};
