+++ src/app/mypage/previous/page.tsx
@@ -1,13 +1,16 @@
 // components
 import TopBar from "@common/TopBar";
 import SavedCats from "./_components/SavedCats";
+import { userServerApi } from "@api/userServer";
 
-export default function Previous() {
+export default async function Previous() {
+  const cupcatData = await userServerApi.getCupcats();
+
   return (
     <div className="flex w-full h-full flex-col">
       <TopBar title="지나간 컵캣" />
       <div className="flex-1 overflow-auto bg-Blue_Grey">
-        <SavedCats />
+        <SavedCats cupcatData={cupcatData} />
       </div>
     </div>
   );
